feat(app.py): 어드민이 만든 메모를 다른 사용자가 바꿀 수 없도록 수정했다.

+++ app.py
@@ -97,7 +97,12 @@ def addMemo():
 @app.route('/memo/delete', methods=['GET'])
 def delMemo():
     id = request.args.get('id')
-    memoRepo.removeMemo(id)
+    memos = memoRepo.getMemos({"_id": id})
+    if (memos[0]['user'] == "admin@admin"):
+        if (session.get('userId') == "admin@admin"):
+            memoRepo.removeMemo(id)
+    else:
+        memoRepo.removeMemo(id)
     return redirect('/')
 
 @app.route('/memo/modify', methods=['POST'])
